chore: simplify filesystem and fix navigation gradient color bug
