+++ src/components/DisplayText.tsx
