+++ src/components/Variables.ts
