--- src/components/containers/index.ts
@@ -1,2 +1,0 @@
-export * from './PageContent';
-export * from './PageView';
