+++ src/containers/StarsBackground.tsx
