--- src/containers/index.ts
@@ -1,3 +1,0 @@
-export * from './Footer';
-export * from './Navbar';
-export * from './StarsBackground';
