+++ src/pages/ethos/index.tsx
@@ -1,16 +1,45 @@
 import styled from 'styled-components';
 
-import { PageContent } from '~/components/containers';
+import { Ball, MOBILE_MAX_WIDTH, SPACING_192, SPACING_512, SPACING_700 } from '~/components';
+import { PageContent } from '~/containers/PageContent';
+import Meatadata from '~/components/Meatadata';
 import ApproachSection from './ApproachSection';
-import { Ball, MOBILE_MAX_WIDTH, SPACING_192, SPACING_512, SPACING_700 } from '~/components/common';
 import TextSection from './TextSection';
+
 import Cone from '~/assets/cone.png';
 import HoopTop from '~/assets/hoop-top.png';
 import HoopBottom from '~/assets/hoop-bottom.png';
 import VIDEO_CHROME from '~/assets/videos/ethos.webm';
 import VIDEO_SAFARI from '~/assets/videos/ethos.mp4';
 import { TitleContainer } from '../landing/HeroSection';
-import Meatadata from '~/components/common/Meatadata';
+
+export default function Ethos() {
+  return (
+    <>
+      <Meatadata title='Ethos' />
+      <PageContent>
+        <HeroDivider>
+          <STitleContainer>
+            <video autoPlay loop muted playsInline>
+              <source src={VIDEO_CHROME} type='video/webm' />
+              <source src={VIDEO_SAFARI} type='video/mp4; codecs="hvc1"' />
+            </video>
+          </STitleContainer>
+        </HeroDivider>
+        <BackgroundContainer>
+          <SCone src={HoopTop.src} alt='starts background' />
+          <SCone src={Cone.src} alt='starts background' />
+          <SBall />
+          <SHoop src={HoopBottom.src} alt='starts background' />
+        </BackgroundContainer>
+
+        <TextSection />
+
+        <StyledApproachSection />
+      </PageContent>
+    </>
+  );
+}
 
 const StyledApproachSection = styled(ApproachSection)`
   padding: '3rem 0';
@@ -38,10 +67,6 @@ const SBall = styled(Ball)`
     top: 14rem;
     width: 6rem;
   }
-`;
-
-const MobileTitleContainer = styled.div`
-  transform: rotate(3deg);
 `;
 
 const BackgroundContainer = styled.div`
@@ -77,31 +102,3 @@ const STitleContainer = styled(TitleContainer)`
     width: 80%;
   }
 `;
-
-export default function Ethos() {
-  return (
-    <>
-      <Meatadata title='Ethos' />
-      <PageContent>
-        <HeroDivider>
-          <STitleContainer>
-            <video autoPlay loop muted playsInline>
-              <source src={VIDEO_CHROME} type='video/webm' />
-              <source src={VIDEO_SAFARI} type='video/mp4; codecs="hvc1"' />
-            </video>
-          </STitleContainer>
-        </HeroDivider>
-        <BackgroundContainer>
-          <SCone src={HoopTop.src} alt='starts background' />
-          <SCone src={Cone.src} alt='starts background' />
-          <SBall />
-          <SHoop src={HoopBottom.src} alt='starts background' />
-        </BackgroundContainer>
-
-        <TextSection />
-
-        <StyledApproachSection />
-      </PageContent>
-    </>
-  );
-}
