+++ src/pages/landing/index.tsx
@@ -1,6 +1,6 @@
 import HeroSection from './HeroSection';
 import LandingContent from './ContentSection';
-import Meatadata from '~/components/common/Meatadata';
+import Meatadata from '~/components/Meatadata';
 
 export default function Landing() {
   return (
